Refactor vote buttons to move out the button styling and fix zero vote count
