+++ src/feedback/talk/text/TalkVoteTextResult.js
